refactor(primitive): merge primitives.py and primitive_base.py
